move: move several actions
