+++ ide/src/main/java/org/aya/ide/action/ComputeTerm.java
@@ -1,6 +1,6 @@
 // Copyright (c) 2020-2022 Tesla (Yinsen) Zhang.
 // Use of this source code is governed by the MIT license that can be found in the LICENSE.md file.
-package org.aya.lsp.actions;
+package org.aya.ide.action;
 
 import org.aya.cli.library.source.LibrarySource;
 import org.aya.concrete.Expr;
