+++ ide/src/main/java/org/aya/ide/action/GotoDefinition.java
@@ -1,6 +1,6 @@
 // Copyright (c) 2020-2022 Tesla (Yinsen) Zhang.
 // Use of this source code is governed by the MIT license that can be found in the LICENSE.md file.
-package org.aya.lsp.actions;
+package org.aya.ide.action;
 
 import kala.collection.SeqView;
 import org.aya.cli.library.source.LibraryOwner;
@@ -9,35 +9,17 @@ import org.aya.concrete.stmt.GeneralizedVar;
 import org.aya.ide.Resolver;
 import org.aya.ide.util.ModuleVar;
 import org.aya.ide.util.XY;
-import org.aya.lsp.utils.*;
 import org.aya.ref.DefVar;
 import org.aya.ref.LocalVar;
 import org.aya.util.error.SourcePos;
 import org.aya.util.error.WithPos;
-import org.javacs.lsp.GenericLocation;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
-
-import java.util.List;
-import java.util.stream.Collectors;
 
 /**
  * @author ice1000, kiva
  */
 public interface GotoDefinition {
-  static @NotNull List<GenericLocation> invoke(
-    @NotNull LibrarySource source,
-    @NotNull SeqView<LibraryOwner> libraries, XY xy
-  ) {
-    return findDefs(source, libraries, xy).mapNotNull(pos -> {
-      var from = pos.sourcePos();
-      var to = pos.data();
-      var res = LspRange.toLoc(from, to);
-      if (res != null) Log.d("Resolved: %s in %s", to, res.targetUri);
-      return res;
-    }).collect(Collectors.toList());
-  }
-
   static @NotNull SeqView<WithPos<SourcePos>> findDefs(
     @NotNull LibrarySource source,
     @NotNull SeqView<LibraryOwner> libraries, XY xy
