+++ lsp/src/main/java/org/aya/lsp/actions/LensMaker.java
@@ -10,6 +10,7 @@ import org.aya.cli.library.source.LibraryOwner;
 import org.aya.cli.library.source.LibrarySource;
 import org.aya.concrete.stmt.Decl;
 import org.aya.concrete.stmt.Stmt;
+import org.aya.ide.action.FindReferences;
 import org.aya.ide.syntax.SyntaxDeclAction;
 import org.aya.lsp.utils.LspRange;
 import org.aya.ide.Resolver;
