+++ lsp/src/main/java/org/aya/lsp/actions/ProjectSymbol.java
@@ -9,6 +9,7 @@ import org.aya.cli.library.source.LibraryOwner;
 import org.aya.cli.library.source.LibrarySource;
 import org.aya.concrete.stmt.Decl;
 import org.aya.concrete.stmt.Stmt;
+import org.aya.ide.action.ComputeSignature;
 import org.aya.ide.syntax.SyntaxDeclAction;
 import org.aya.lsp.utils.LspRange;
 import org.aya.ide.Resolver;
